+++ src/UI/IconContainer/IconContainer.module.scss
@@ -4,6 +4,11 @@
   padding: 24px;
   @include mixines.border(1px, var(--element-color));
   border-bottom: 0;
+  transition: all 0.3s;
+
+  &:hover {
+    border-color: var(--main-color);
+  }
 }
 
 .icon {
