+++ src/UI/NavLink/NavigationLink.module.scss
@@ -1,7 +1,6 @@
 .link {
   color: var(--text-color);
   text-decoration: none;
-  text-transform: uppercase;
 }
 
 .active {
