+++ src/UI/NavLink/NavigationLink.tsx
@@ -1,22 +1,24 @@
 import { NavLink, NavLinkRenderProps } from 'react-router-dom';
 
+import s from './NavigationLink.module.scss';
+
 type Props = {
   to: string;
-  link: string | React.ReactNode;
+  label: string | React.ReactNode;
   className?: string | ((props: NavLinkRenderProps) => string | undefined);
   handleClick?: () => void;
 };
 
 const NavigationLink: React.FC<Props> = (props) => {
-  const { to, link, className, handleClick } = props;
+  const { to, label, className, handleClick } = props;
 
   return (
     <NavLink
       to={to}
-      className={className ? className : ''}
+      className={className ? className : s.link}
       onClick={handleClick}
     >
-      {link}
+      {label}
     </NavLink>
   );
 };
